Update seed with new models

+++ server/lib/database/live/pack_scene.ex
@@ -5,7 +5,7 @@ defmodule Database.Live.PackScene do
     belongs_to :pack, Pack
     belongs_to :scene, Scene
 
-    field :order, :decimal
+    field :order, :integer
 
     timestamps()
   end
